:memo: Better TypeLists
