+++ src/Game.inl.hxx
@@ -1,15 +1,12 @@
 
 template <typename ArgumentType> void Game::UIVisitor::operator()(ArgumentType &&event)
 {
-  static_assert(
-      sizeof(ArgumentType) < 0,
-                "UIVisitor does not know how to handle this event. You must specialize the functor");
+  static_assert(sizeof(ArgumentType) < 0, "UIVisitor does not know how to handle this event. You must specialize the functor");
 }
-
 
 #ifdef USE_AUDIO
 template <typename AudioEventType>
-EnableIf<ContainsType<AudioEvents, AudioEventType>, void> Game::GameVisitor::operator()(AudioEventType &&event)
+EnableIf<AudioEvents::ContainsType<AudioEventType>, void> Game::GameVisitor::operator()(AudioEventType &&event)
 {
   GetService<AudioMixer>().handleEvent(std::move(event));
 }
@@ -17,7 +14,5 @@ EnableIf<ContainsType<AudioEvents, AudioEventType>, void> Game::GameVisitor::ope
 
 template <typename ArgumentType> void Game::GameVisitor::operator()(const ArgumentType &&event)
 {
-  static_assert(
-      sizeof(ArgumentType) < 0,
-                "GameVisitor does not know how to handle this event. You must specialize the functor");
+  static_assert(sizeof(ArgumentType) < 0, "GameVisitor does not know how to handle this event. You must specialize the functor");
 }
